requirements.txt
Flask==2.1.1
flask-cors==3.1.1
notion-client==0.1.0
pyperclip==1.8.2
python-dotenv==0.19.1
